renamed common to gtime
