+++ go/genkit/schemas.config
@@ -1,7 +1,7 @@
 # This file holds configuration for the genkit-schema.json file
 # generated by the npm export:schemas script.
 
-genkit							import github.com/firebase/genkit/go/common
+genkit							import github.com/firebase/genkit/go/gtime
 
 # DocumentData type was hand-written.
 DocumentData						omit
@@ -69,8 +69,8 @@ FlowInvokeEnvelopeMessageState		omit
 Operation						omit
 
 FlowStateExecution				name FlowExecution
-FlowStateExecution.startTime	type common.Milliseconds
-FlowStateExecution.endTime		type common.Milliseconds
+FlowStateExecution.startTime	type gtime.Milliseconds
+FlowStateExecution.endTime		type gtime.Milliseconds
 
 FlowError						name flowError
