+++ go/genkit/trace_store_exporter_test.go
@@ -18,7 +18,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/firebase/genkit/go/common"
+	"github.com/firebase/genkit/go/gtime"
 	"github.com/firebase/genkit/go/gtrace"
 	"github.com/google/go-cmp/cmp"
 	"go.opentelemetry.io/otel/attribute"
@@ -91,11 +91,11 @@ func TestConvertSpan(t *testing.T) {
 		SpanID:       spanID1,
 		ParentSpanID: spanID2,
 		SpanKind:     "INTERNAL",
-		StartTime:    common.Milliseconds(1e3),
-		EndTime:      common.Milliseconds(2e3),
+		StartTime:    gtime.Milliseconds(1e3),
+		EndTime:      gtime.Milliseconds(2e3),
 		Attributes:   map[string]any{"k": "v"},
 		TimeEvents: gtrace.TimeEvents{TimeEvent: []gtrace.TimeEvent{{
-			Time: common.Milliseconds(3e3),
+			Time: gtime.Milliseconds(3e3),
 			Annotation: gtrace.Annotation{
 				Attributes:  map[string]any{"k2": "v2"},
 				Description: "ename",
