+++ go/gtrace/store.go
@@ -19,7 +19,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/firebase/genkit/go/common"
+	"github.com/firebase/genkit/go/gtime"
 )
 
 // A Store stores trace information.
@@ -62,8 +62,8 @@ type Query struct {
 type Data struct {
 	TraceID     string               `json:"traceId"`
 	DisplayName string               `json:"displayName"`
-	StartTime   common.Milliseconds  `json:"startTime"`
-	EndTime     common.Milliseconds  `json:"endTime"`
+	StartTime   gtime.Milliseconds   `json:"startTime"`
+	EndTime     gtime.Milliseconds   `json:"endTime"`
 	Spans       map[string]*SpanData `json:"spans"`
 }
 
@@ -76,8 +76,8 @@ type SpanData struct {
 	SpanID                 string                 `json:"spanId"`
 	TraceID                string                 `json:"traceId,omitempty"`
 	ParentSpanID           string                 `json:"parentSpanId,omitempty"`
-	StartTime              common.Milliseconds    `json:"startTime"`
-	EndTime                common.Milliseconds    `json:"endTime"`
+	StartTime              gtime.Milliseconds     `json:"startTime"`
+	EndTime                gtime.Milliseconds     `json:"endTime"`
 	Attributes             map[string]any         `json:"attributes,omitempty"`
 	DisplayName            string                 `json:"displayName"`
 	Links                  []*Link                `json:"links,omitempty"`
@@ -98,8 +98,8 @@ type BoolValue struct {
 }
 
 type TimeEvent struct {
-	Time       common.Milliseconds `json:"time,omitempty"`
-	Annotation Annotation          `json:"annotation,omitempty"`
+	Time       gtime.Milliseconds `json:"time,omitempty"`
+	Annotation Annotation         `json:"annotation,omitempty"`
 }
 
 type Annotation struct {
